split big file and delete unused file
