+++ lib/services/places/places_notifier.dart
@@ -1,0 +1,34 @@
+/// Global notifier for places data changes.
+///
+// Time-stamp: <2026-01-07 Miduo>
+///
+/// Copyright (C) 2025, Software Innovation Institute, ANU.
+///
+/// Licensed under the GNU General Public License, Version 3 (the "License").
+///
+/// License: https://opensource.org/license/gpl-3-0.
+//
+// This program is free software: you can redistribute it and/or modify it under
+// the terms of the GNU General Public License as published by the Free Software
+// Foundation, either version 3 of the License, or (at your option) any later
+// version.
+//
+// This program is distributed in the hope that it will be useful, but WITHOUT
+// ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+// FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
+// details.
+//
+// You should have received a copy of the GNU General Public License along with
+// this program.  If not, see <https://opensource.org/license/gpl-3-0>.
+///
+/// Authors: Graham Williams, Miduo
+
+library;
+
+import 'package:flutter/material.dart';
+
+/// Global notifier that increments whenever places data changes.
+///
+/// Listeners can watch this to rebuild or refresh whenever any place is added,
+/// updated, deleted, or imported.
+final placesChangeNotifier = ValueNotifier<int>(0);
