+++ lib/services/places_service.dart
@@ -27,538 +27,174 @@
 
 library;
 
-import 'dart:convert';
-
 import 'package:flutter/material.dart';
 
-import 'package:solidpod/solidpod.dart';
-
-import 'package:geopod/constants/example_places_data.dart';
 import 'package:geopod/models/place.dart';
-import 'package:geopod/services/geocoding_service.dart';
-import 'package:geopod/services/places/encrypted_places_service.dart';
-import 'package:geopod/services/places/places_cache_manager.dart';
-import 'package:geopod/services/places/places_cache_persistence.dart';
+import 'package:geopod/services/places/places_cache_service.dart';
+import 'package:geopod/services/places/places_fetch_service.dart';
 import 'package:geopod/services/places/places_import_export.dart';
-import 'package:geopod/services/places/places_pod_file.dart';
-import 'package:geopod/services/pod/pod_directory_service.dart';
+import 'package:geopod/services/places/places_write_service.dart';
 
 export 'package:geopod/models/place.dart';
 export 'package:geopod/services/places/encrypted_places_service.dart'
     show EncryptedPlacesService;
 export 'package:geopod/services/places/places_cache_manager.dart';
 export 'package:geopod/services/places/places_import_export.dart';
+export 'package:geopod/services/places/places_notifier.dart'
+    show placesChangeNotifier;
 
-final placesChangeNotifier = ValueNotifier<int>(0);
+/// Facade that coordinates [PlacesFetchService], [PlacesCacheService], and
+/// [PlacesWriteService].
+///
+/// All public static methods preserve the original API so callers require no
+/// changes.  Implementation details live in the sub-services under
+/// `lib/services/places/`.
 
 class PlacesService {
-  /// Cached local places (lazily initialized from compiled constants).
-  static List<Place>? _cachedLocalPlaces;
+  //  Fetch
 
-  /// Get local example places synchronously.
+  /// Get local example places synchronously (compiled into the binary).
 
-  static List<Place> getLocalPlacesSync() {
-    _cachedLocalPlaces ??= kExamplePlacesData
-        .map((json) => Place.fromJson(json, isLocalSource: true))
-        .toList();
-    return _cachedLocalPlaces!;
-  }
+  static List<Place> getLocalPlacesSync() =>
+      PlacesFetchService.getLocalPlacesSync();
 
-  /// Load local example places (async wrapper for API compatibility).
-  /// NOTE: Prefer using getLocalPlacesSync() directly as local places
-  /// are now compiled into the app and don't require async loading.
+  /// Load local example places.
+  /// NOTE: Prefer [getLocalPlacesSync]  local data is compiled into the app.
 
   @Deprecated('Use getLocalPlacesSync() instead - local data is compiled in')
-  static Future<List<Place>> loadLocalPlaces() async => getLocalPlacesSync();
+  static Future<List<Place>> loadLocalPlaces() async =>
+      PlacesFetchService.getLocalPlacesSync();
+
+  /// Fetch all places: local, Pod, and optionally encrypted.
 
   static Future<List<Place>> fetchPlaces({
     bool forceRefresh = false,
     bool includeEncrypted = false,
-  }) async {
-    final cm = PlacesCacheManager();
-    if (!forceRefresh) {
-      final c = cm.allPlaces;
-      if (c != null) {
-        // If cached but need encrypted and none present, try fast-merge from
-        // EncryptedPlacesService in-memory cache before any network call.
-        if (includeEncrypted && !c.any((p) => p.isEncrypted)) {
-          final cachedEnc = EncryptedPlacesService.getCachedEncryptedPlaces();
-          if (cachedEnc != null && cachedEnc.isNotEmpty) {
-            final merged = [...c, ...cachedEnc];
-            cm.cacheAllPlaces(merged);
-            return merged;
-          }
-          // No encrypted cache — fall through to network fetch.
-        } else {
-          return c;
-        }
-      }
-    }
+  }) => PlacesFetchService.fetchPlaces(
+    forceRefresh: forceRefresh,
+    includeEncrypted: includeEncrypted,
+  );
 
-    // Local places are synchronous (compiled into binary) - get them immediately.
-    final localPlaces = getLocalPlacesSync();
-
-    // Fetch network data in parallel for better performance.
-    final results = await Future.wait([
-      fetchPodPlaces(forceRefresh: forceRefresh),
-      includeEncrypted
-          ? fetchEncryptedPlaces(forceRefresh: forceRefresh)
-          : Future.value(<Place>[]),
-    ]);
-
-    final podPlaces = results[0];
-    final encryptedPlaces = results[1];
-
-    final all = <Place>[...podPlaces, ...encryptedPlaces, ...localPlaces];
-    cm.cacheAllPlaces(all);
-    return all;
-  }
-
-  /// Fetch encrypted places from Pod.
-  /// Returns empty list if not logged in or no security key available.
-  /// NOTE: Will not prompt for security key - use EncryptedPlacesService
-  /// directly if you need to prompt the user.
+  /// Fetch encrypted places from the Pod without prompting for a security key.
 
   static Future<List<Place>> fetchEncryptedPlaces({
     bool forceRefresh = false,
-  }) async {
-    try {
-      if (!authStateNotifier.value) return [];
+  }) => PlacesFetchService.fetchEncryptedPlaces(forceRefresh: forceRefresh);
 
-      // Fast path: data already in memory — skip the async key check entirely.
-      if (!forceRefresh && EncryptedPlacesService.hasLoadedEncryptedPlaces) {
-        return EncryptedPlacesService.getCachedEncryptedPlaces()!;
-      }
+  /// Fetch places stored in the user's Solid Pod.
 
-      // Check if security key is available - don't try to load if not.
-      final hasKey = await EncryptedPlacesService.isSecurityKeyAvailable();
-      if (!hasKey) {
-        return [];
-      }
+  static Future<List<Place>> fetchPodPlaces({bool forceRefresh = false}) =>
+      PlacesFetchService.fetchPodPlaces(forceRefresh: forceRefresh);
 
-      final encPlaces = await EncryptedPlacesService.fetchEncryptedPlaces(
-        forceRefresh: forceRefresh,
-      );
-      return encPlaces;
-    } catch (e) {
-      debugPrint('PlacesService.fetchEncryptedPlaces error: $e');
-      return [];
-    }
-  }
+  //  Cache
 
-  static Future<List<Place>> fetchPodPlaces({bool forceRefresh = false}) async {
-    final places = <Place>[];
-    final cm = PlacesCacheManager();
-    try {
-      if (!authStateNotifier.value) return places;
-      if (!forceRefresh) {
-        final mc = cm.podPlaces;
-        if (mc != null) {
-          _refreshPodPlacesInBackground();
-          return mc;
-        }
-      }
-      if (!forceRefresh) {
-        final c = await PlacesCachePersistence.getCachedPodPlaces();
-        if (c != null) {
-          cm.cachePodPlaces(c);
-          _refreshPodPlacesInBackground();
-          return c;
-        }
-      }
-      final content = await readPlacesJsonFile();
-      if (content == null || content.trim().isEmpty) return places;
-      final decoded = jsonDecode(content);
-      if (decoded is List) {
-        for (final i in decoded) {
-          if (i is Map<String, dynamic>) {
-            try {
-              places.add(Place.fromJson(i, isLocalSource: false));
-            } catch (_) {}
-          }
-        }
-      } else if (decoded is Map<String, dynamic>) {
-        try {
-          places.add(Place.fromJson(decoded, isLocalSource: false));
-        } catch (_) {}
-      }
-      places.sort((a, b) => b.timestamp.compareTo(a.timestamp));
-      await PlacesCachePersistence.cachePodPlaces(content);
-      cm.cachePodPlaces(places);
-    } catch (_) {}
-    return places;
-  }
+  /// Clears all caches (in-memory, persisted, and encrypted session state).
 
-  static void _refreshPodPlacesInBackground() {
-    Future(() async {
-      try {
-        if (!authStateNotifier.value) return;
-        final c = await readPlacesJsonFile();
-        if (c != null && c.trim().isNotEmpty) {
-          await PlacesCachePersistence.cachePodPlaces(c);
-        }
-      } catch (_) {}
-    });
-  }
+  static Future<void> clearCache() => PlacesCacheService.clearCache();
 
-  static Future<void> clearCache() async {
-    try {
-      PlacesCacheManager().clearCache();
-      // Both are independent — run in parallel.
-      await Future.wait([
-        PlacesCachePersistence.clearPodPlacesCache(),
-        EncryptedPlacesService.resetSessionState(),
-      ]);
-    } catch (_) {}
-  }
+  /// Clears only the Pod (non-encrypted) portion of the cache.
 
-  static Future<void> clearPodCacheOnly() async {
-    try {
-      PlacesCacheManager().clearPodCacheOnly();
-      await PlacesCachePersistence.clearPodPlacesCache();
-    } catch (_) {}
-  }
+  static Future<void> clearPodCacheOnly() =>
+      PlacesCacheService.clearPodCacheOnly();
 
-  static Future<List<Place>> refreshPodDataOnly() async {
-    final cm = PlacesCacheManager();
+  /// Forces a fresh Pod fetch and returns merged Pod + local places.
 
-    // Local places are synchronous (compiled into binary)
-    final local = getLocalPlacesSync();
-    await clearPodCacheOnly();
-    final pod = await fetchPodPlaces(forceRefresh: true);
-    final all = <Place>[...pod, ...local];
-    cm.cacheAllPlaces(all);
-    return all;
-  }
+  static Future<List<Place>> refreshPodDataOnly() =>
+      PlacesCacheService.refreshPodDataOnly();
+
+  //  Write / Mutate
+
+  /// Adds [place] to the user's Pod and updates all caches.
 
   static Future<bool> addPlace(
     Place place,
     BuildContext context,
     Widget returnWidget,
-  ) async {
-    try {
-      if (!authStateNotifier.value) return false;
-      final cm = PlacesCacheManager();
-      var existing = cm.podPlaces ?? await fetchPodPlaces();
-      final updated = List<Place>.from(existing)..insert(0, place);
+  ) => PlacesWriteService.addPlace(place, context, returnWidget);
 
-      // Write main file first to ensure it succeeds.
-      final mainSuccess = await writePlacesJsonFile(
-        jsonEncode(updated.map((p) => p.toJson()).toList()),
-      );
-
-      if (mainSuccess) {
-        // Fire-and-forget individual file (non-critical — main file is the
-        // source of truth and is already written above).
-        writeIndividualPlaceFile(place);
-
-        // Surgically insert into caches — avoids a full clear + re-fetch and
-        // preserves the encrypted places cache which is unrelated to this write.
-        final newJson = jsonEncode(updated.map((p) => p.toJson()).toList());
-        cm.insertPlaceIntoCache(place);
-        cm.cachePodPlaces(updated);
-        // Persist in background — does not block the UI notification.
-        PlacesCachePersistence.cachePodPlaces(newJson);
-        placesChangeNotifier.value++;
-
-        // Clear directory cache completely to force refresh.
-
-        PodDirectoryService.clearCache();
-        PodDirectoryService.notifyChange();
-      }
-      return mainSuccess;
-    } catch (e) {
-      debugPrint('Error in addPlace: $e');
-      return false;
-    }
-  }
+  /// Deletes the place identified by [placeId] from the user's Pod.
 
   static Future<bool> deletePlace(
     String placeId,
     BuildContext context,
     Widget returnWidget,
-  ) async {
-    try {
-      if (!authStateNotifier.value) return false;
-      final cm = PlacesCacheManager();
-      var existing = cm.podPlaces ?? await fetchPodPlaces();
-      final updated = List<Place>.from(existing)
-        ..removeWhere((p) => p.id == placeId);
+  ) => PlacesWriteService.deletePlace(placeId, context, returnWidget);
 
-      // Delete individual file and update main file in parallel.
-      final results = await Future.wait([
-        writePlacesJsonFile(
-          jsonEncode(updated.map((p) => p.toJson()).toList()),
-        ),
-        deleteIndividualPlaceFile(placeId),
-      ]);
-      final success = results[0];
+  /// Deletes a place, routing to encrypted or regular storage automatically.
 
-      if (success) {
-        // Surgically remove from in-memory caches and refresh pod persistence.
-        cm.removePlaceFromCache(placeId);
-        await PlacesCachePersistence.cachePodPlaces(
-          jsonEncode(updated.map((p) => p.toJson()).toList()),
-        );
-        cm.cachePodPlaces(updated);
-        placesChangeNotifier.value++;
-
-        // Invalidate directory cache and notify file browser.
-
-        PodDirectoryService.invalidateCache('data/places');
-        PodDirectoryService.notifyChange();
-      }
-      return success;
-    } catch (_) {
-      return false;
-    }
-  }
-
-  static Future<bool> exportPlaces(List<Place> places) =>
-      PlacesImportExport.exportPlaces(places);
-  static Future<ImportResult> importPlaces() =>
-      PlacesImportExport.importPlaces();
-
-  static Future<bool> mergeImportedPlaces(
-    List<Place> imported,
-    BuildContext context,
-    Widget returnWidget, {
-    void Function(int, int)? onProgress,
-  }) async {
-    try {
-      if (!authStateNotifier.value) return false;
-      final existing = await fetchPodPlaces();
-      final ids = existing.map((p) => p.id).toSet();
-      final newPlaces = imported.where((p) => !ids.contains(p.id)).toList();
-      if (newPlaces.isEmpty && imported.isNotEmpty) return true;
-
-      // Fetch all addresses in parallel instead of sequentially.
-      onProgress?.call(0, newPlaces.length);
-      final addresses = await Future.wait(
-        newPlaces.map((p) => GeocodingService.getAddress(p.lat, p.lng)),
-      );
-      onProgress?.call(newPlaces.length, newPlaces.length);
-
-      final withAddr = List<Place>.generate(newPlaces.length, (i) {
-        final p = newPlaces[i];
-        return Place(
-          id: p.id,
-          lat: p.lat,
-          lng: p.lng,
-          note: p.note,
-          timestamp: p.timestamp,
-          address: addresses[i],
-          isLocal: false,
-        );
-      });
-      final merged = [...withAddr, ...existing];
-
-      // Write main file first.
-      final success = await writePlacesJsonFile(
-        jsonEncode(merged.map((p) => p.toJson()).toList()),
-      );
-
-      if (success) {
-        // Write individual files for new places in parallel.
-        await Future.wait(withAddr.map((p) => writeIndividualPlaceFile(p)));
-        await clearCache();
-        placesChangeNotifier.value++;
-
-        // Invalidate directory cache and notify file browser.
-
-        PodDirectoryService.invalidateCache('data/places');
-        PodDirectoryService.notifyChange();
-      }
-      return success;
-    } catch (_) {
-      return false;
-    }
-  }
-
-  static Future<bool> clearAllPlaces(
+  static Future<bool> deletePlaceByPlace(
+    Place place,
     BuildContext context,
     Widget returnWidget,
-  ) async {
-    try {
-      if (!authStateNotifier.value) return false;
+  ) => PlacesWriteService.deletePlaceByPlace(place, context, returnWidget);
 
-      // Get all existing place IDs before clearing.
-      final existing = await fetchPodPlaces();
-      final placeIds = existing.map((p) => p.id).toList();
-      // Snapshot before async write so we know whether to wipe encrypted.
-      final hadEncryptedPlaces =
-          EncryptedPlacesService.hasLoadedEncryptedPlaces;
+  /// Deletes a place by its individual file path (e.g. from the file browser).
 
-      final success = await writePlacesJsonFile('[]');
-      if (success) {
-        // Delete all individual place files.
-        await deleteAllIndividualPlaceFiles(placeIds);
+  static Future<bool> deletePlaceByFilePath(
+    String filePath,
+    BuildContext context,
+    Widget returnWidget,
+  ) =>
+      PlacesWriteService.deletePlaceByFilePath(filePath, context, returnWidget);
 
-        if (hadEncryptedPlaces) {
-          // Clear main/pod cache but keep encrypted session state alive so
-          // writeEncryptedPlaces can skip the security key re-check.
-          PlacesCacheManager().clearCache();
-          await PlacesCachePersistence.clearPodPlacesCache();
-          // Erase encrypted places from Pod. writeEncryptedPlaces updates
-          // the encrypted cache and fires placesChangeNotifier.
-          await EncryptedPlacesService.writeEncryptedPlaces(
-            [],
-            context,
-            returnWidget,
-          );
-        } else {
-          await clearCache();
-          placesChangeNotifier.value++;
-        }
-
-        // Invalidate directory cache and notify file browser.
-
-        PodDirectoryService.invalidateCache('data/places');
-        PodDirectoryService.notifyChange();
-      }
-      return success;
-    } catch (_) {
-      return false;
-    }
-  }
+  /// Updates [updated] in the Pod, optionally re-geocoding its address.
 
   static Future<bool> updatePlace(
     Place updated,
     BuildContext context,
     Widget returnWidget, {
     bool coordinatesChanged = false,
-  }) async {
-    try {
-      if (!authStateNotifier.value) return false;
+  }) => PlacesWriteService.updatePlace(
+    updated,
+    context,
+    returnWidget,
+    coordinatesChanged: coordinatesChanged,
+  );
 
-      // Route encrypted places to their dedicated storage.
-      if (updated.isEncrypted) {
-        var toSave = updated;
-        if (coordinatesChanged) {
-          final addr = await GeocodingService.getAddress(
-            updated.lat,
-            updated.lng,
-          );
-          toSave = Place(
-            id: updated.id,
-            lat: updated.lat,
-            lng: updated.lng,
-            note: updated.note,
-            timestamp: updated.timestamp,
-            address: addr,
-            isLocal: false,
-            isEncrypted: true,
-          );
-        }
-        // Surgically update main cache BEFORE the write so that the
-        // placesChangeNotifier fired inside writeEncryptedPlaces hits an
-        // already-correct allPlaces cache — no revert, no extra network fetch.
-        PlacesCacheManager().updatePlaceInCache(toSave);
-        return EncryptedPlacesService.updateEncryptedPlace(
-          toSave,
-          context,
-          returnWidget,
-        );
-      }
+  /// Removes all places from the user's Pod (regular and encrypted).
 
-      final cm = PlacesCacheManager();
-      final existing = cm.podPlaces ?? await fetchPodPlaces();
-      final list = List<Place>.from(existing);
-      final i = list.indexWhere((p) => p.id == updated.id);
-      var toSave = updated;
-
-      if (i == -1) {
-        list.insert(0, updated);
-      } else {
-        if (coordinatesChanged) {
-          final addr = await GeocodingService.getAddress(
-            updated.lat,
-            updated.lng,
-          );
-          toSave = Place(
-            id: updated.id,
-            lat: updated.lat,
-            lng: updated.lng,
-            note: updated.note,
-            timestamp: updated.timestamp,
-            address: addr,
-            isLocal: false,
-          );
-        }
-        list[i] = toSave;
-      }
-
-      // Update both main file and individual file in parallel.
-      final newJson = jsonEncode(list.map((p) => p.toJson()).toList());
-      final results = await Future.wait([
-        writePlacesJsonFile(newJson),
-        writeIndividualPlaceFile(toSave),
-      ]);
-      final success = results[0];
-
-      if (success) {
-        // Surgically update in-memory caches — no full reload needed.
-        cm.updatePlaceInCache(toSave);
-        await PlacesCachePersistence.cachePodPlaces(newJson);
-        cm.cachePodPlaces(list);
-        placesChangeNotifier.value++;
-
-        // Invalidate directory cache and notify file browser.
-
-        PodDirectoryService.invalidateCache('data/places');
-        PodDirectoryService.notifyChange();
-      }
-      return success;
-    } catch (_) {
-      return false;
-    }
-  }
-
-  /// Deletes a place routing to the correct storage based on [place.isEncrypted].
-
-  static Future<bool> deletePlaceByPlace(
-    Place place,
+  static Future<bool> clearAllPlaces(
     BuildContext context,
     Widget returnWidget,
-  ) async {
-    if (place.isEncrypted) {
-      final success = await EncryptedPlacesService.deleteEncryptedPlace(
-        place.id,
-        context,
-        returnWidget,
-      );
-      if (success) placesChangeNotifier.value++;
-      return success;
-    }
-    return deletePlace(place.id, context, returnWidget);
-  }
+  ) => PlacesWriteService.clearAllPlaces(context, returnWidget);
 
-  /// Delete a place by its individual file path.
-  /// This is called when a user deletes place_xxx.json from the file browser.
-  /// It will also remove the place from the main places.json file.
+  //  Import / Export
 
-  static Future<bool> deletePlaceByFilePath(
-    String filePath,
+  /// Exports [places] to a local file.
+
+  static Future<bool> exportPlaces(List<Place> places) =>
+      PlacesWriteService.exportPlaces(places);
+
+  /// Opens a file picker and imports places from the selected file.
+
+  static Future<ImportResult> importPlaces() =>
+      PlacesWriteService.importPlaces();
+
+  /// Merges [imported] places into the user's Pod, skipping duplicates.
+
+  static Future<bool> mergeImportedPlaces(
+    List<Place> imported,
     BuildContext context,
-    Widget returnWidget,
-  ) async {
-    // Extract place ID from file path like "place_abc123.json"
-    final fileName = filePath.split('/').last;
-    final match = RegExp(r'^place_(.+)\.json$').firstMatch(fileName);
-    if (match == null) return false;
+    Widget returnWidget, {
+    void Function(int, int)? onProgress,
+  }) => PlacesWriteService.mergeImportedPlaces(
+    imported,
+    context,
+    returnWidget,
+    onProgress: onProgress,
+  );
 
-    final placeId = match.group(1)!;
-    return deletePlace(placeId, context, returnWidget);
-  }
+  //  Utilities
 
-  /// Check if a file path is a places.json file.
+  /// Returns `true` if [filePath] points to the main `places.json` file.
 
   static bool isMainPlacesFile(String filePath) {
     return filePath.endsWith('/places.json') ||
         filePath.endsWith('\\places.json');
   }
 
-  /// Check if a file path is an individual place file.
+  /// Returns `true` if [filePath] points to an individual place file
+  /// (`place_<id>.json`).
 
   static bool isIndividualPlaceFile(String filePath) {
     final fileName = filePath.split('/').last.split('\\').last;
